refactor: move domain classes to separate file
