+++ txstratum/api.py
@@ -13,7 +13,7 @@ from structlog import get_logger
 
 import txstratum.time
 from txstratum.exceptions import JobAlreadyExists, NewJobRefused
-from txstratum.healthcheck import HealthCheck
+from txstratum.healthcheck.healthcheck import HealthCheck
 from txstratum.jobs import JobStatus, TxJob
 from txstratum.middleware import create_middleware_version_check
 from txstratum.utils import tx_or_block_from_bytes
